requirements.txt
PyYAML
numpy
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
soundfile
sounddevice
